Conflicts:
	modules/enterprise/gui/coregui/src/main/java/org/rhq/enterprise/gui/coregui/client/report/AlertDefinitionReportView.java

+++ modules/enterprise/server/jar/src/test/java/org/rhq/enterprise/server/util/ResourceTreeHelper.java
@@ -166,6 +166,10 @@ public class ResourceTreeHelper {
     }
 
     public static void deleteResource(EntityManager entityManager, Resource doomedResource) {
+        if (null == doomedResource) {
+            return;
+        }
+
         for (Availability avail : doomedResource.getAvailability()) {
             entityManager.remove(avail);
         }
